backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
psutil==5.9.8
paramiko==3.4.0
cryptography==42.0.5
gunicorn==21.2.0
whitenoise==6.6.0
django-environ==0.11.2

